src/config-validator/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
ruamel.yaml>=0.18.0
pydantic>=2.9.0
